+++ changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 3.0.4
+
 - Add random command on koth win
 - Fix error with discord webhook on stop
 - Fix problem with placeholder on koth spawn